feat: response gathering and timeout
When the container is running for longer than the specified timeout, the container will be killed. The timeout is triggered once the event for the start has been given. Additionally once the container has been removed. The worker will get the response and process it (also cleaning up the sandbox).